requirements.txt
openai>=1.0.0
langchain>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0 